% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_lapost_grd.R
\name{mapview_lapost_grd}
\alias{mapview_lapost_grd}
\title{mapview_lalpost_grd}
\usage{
mapview_lapost_grd(
  file_grd,
  epsg = 32632,
  at = NULL,
  export = FALSE,
  string_filename = " file_name",
  name_of_map_layer = "layer_name"
)
}
\arguments{
\item{file_grd}{string path to the LAPOST ascii grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{at}{numeric vector defining breakpoints for the visualization,
default value is NULL by using continuous 'pretty' breakpoints, see lattice::levelplot function for details}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview raster object, eventually exported as html and png files
}
\description{
Create a mapview raster object from LAPOST ascii grid file
}
