% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_lapost_grd_contour.R
\name{mapview_lapost_grd_contour}
\alias{mapview_lapost_grd_contour}
\title{mapview_lalpost_grd}
\usage{
mapview_lapost_grd_contour(
  file_grd,
  epsg = 32632,
  levels = NULL,
  string_filename = " file_name",
  name_of_map_layer = "layer_name"
)
}
\arguments{
\item{file_grd}{string path to the LAPOST ascii grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{levels}{vector of contour levels to plot}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from LAPOST ascii grid file and export as html and png files
}
