% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_lapost_grd_surfer.R
\name{mapview_lapost_grd_surfer}
\alias{mapview_lapost_grd_surfer}
\title{Plot LAPOST surfer grid file in mapview}
\usage{
mapview_lapost_grd_surfer(
  file_grd,
  epsg = 32632,
  at = NULL,
  export = FALSE,
  name_of_map_layer = "mapview_layer_name",
  string_filename = "export_filename",
  trans_factor = 1
)
}
\arguments{
\item{file_grd}{string path to the LAPOST grid file in surfer format}

\item{epsg}{number epsg to set crs in the raster object}

\item{at}{numeric vector defining breakpoints for the visualization,
default value is NULL by using continuous 'pretty' breakpoints, see lattice::levelplot function for details}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{name_of_map_layer}{string to name the mapview layer}

\item{string_filename}{string to name the output files}

\item{trans_factor}{number, transformation factor to be applied to calculated values, leave it as default = 1}
}
\value{
mapview raster object, eventually exported as html and png files
}
\description{
Create a mapview object from LAPOST grid file in surfer format.
\emph{Beware: it works ONLY with grid files in surfer format}
}
