% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lapost_grd.R
\name{read_lapost_grd}
\alias{read_lapost_grd}
\title{read_lapost_grd}
\usage{
read_lapost_grd(file_grd, epsg = 32632)
}
\arguments{
\item{file_grd}{string path to LAPOST ascii grid file}

\item{epsg}{number epsg to set crs in the raster object}
}
\value{
raster object
}
\description{
Read LAPOST ascii grid file
}
\keyword{internal}
