% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lapost_grd_surfer.R
\name{read_lapost_grd_surfer}
\alias{read_lapost_grd_surfer}
\title{read_lapost_grd_surfer}
\usage{
read_lapost_grd_surfer(file_grd, epsg = 32632)
}
\arguments{
\item{file_grd}{string path to LAPOST grid file in SURFER ASCII format}

\item{epsg}{number epsg to set crs in the raster object}
}
\value{
raster object
}
\description{
Read LAPOST grid file in SURFER ASCII format
}
\keyword{internal}
