% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_lapost_grd_contour.R
\name{mapview_lapost_grd_contour}
\alias{mapview_lapost_grd_contour}
\title{Plot contour of LAPOST acii grid file in mapview}
\usage{
mapview_lapost_grd_contour(
  file_grd,
  epsg = 32632,
  levels = NULL,
  export = FALSE,
  string_filename = " file_name",
  name_of_map_layer = "layer_name"
)
}
\arguments{
\item{file_grd}{string path to the LAPOST ascii grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{levels}{vector of the levels for the contourplot}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview contourline object, eventually exported as html and png files
}
\description{
Create a mapview contourline object from LAPOST ascii grid file
}
